Add delete alert to delete accessory tap

+++ Task Master/View Controllers/Folders/FolderViewController.swift
@@ -110,14 +110,26 @@ class FolderViewController: UIViewController {
                 let headerDisclosureOption = UICellAccessory.OutlineDisclosureOptions(style: .header)
                 cell.accessories = [
                     .delete(displayed: .whenEditing, actionHandler: {
-                        self.viewModel.deleteFolder(item)
+                        let ac = UIAlertController(title: "Delete Folder", message: "Are you sure you want to delete this folder? Deleting this folder will also delete all associated tasks. This action cannot be undone.", preferredStyle: .alert)
+                        ac.addAction(UIAlertAction(title: "Cancel", style: .default))
+                        ac.addAction(UIAlertAction(title: "Delete", style: .destructive, handler: { [weak self] _ in
+                            guard let self = self else { return }
+                            self.viewModel.deleteFolder(item)
+                        }))
+                        self.present(ac, animated: true)
                     }),
                     .outlineDisclosure(options:headerDisclosureOption)
                 ]
             } else {
                 cell.accessories = [
                     .delete(displayed: .whenEditing, actionHandler: {
-                        self.viewModel.deleteFolder(item)
+                        let ac = UIAlertController(title: "Delete Folder", message: "Are you sure you want to delete this folder? Deleting this folder will also delete all associated tasks. This action cannot be undone.", preferredStyle: .alert)
+                        ac.addAction(UIAlertAction(title: "Cancel", style: .default))
+                        ac.addAction(UIAlertAction(title: "Delete", style: .destructive, handler: { [weak self] _ in
+                            guard let self = self else { return }
+                            self.viewModel.deleteFolder(item)
+                        }))
+                        self.present(ac, animated: true)
                     })
                 ]
             }
